pyproject.toml
[project]
name = "t-inv-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "aio-pika>=9.5.8",
    "aiogram>=3.24.0",
    "alembic>=1.18.4",
    "dishka>=1.7.2",
    "faststream[rabbit]>=0.6.6",
    "llama-index>=0.14.14",
    "llama-index-llms-ollama>=0.9.1",
    "numba>=0.64.0",
    "polars>=1.38.1",
    "psycopg>=3.3.2",
    "pyyaml>=6.0.3",
    "qdrant-client>=1.16.2",
    "redis>=7.1.1",
    "sentence-transformers>=5.2.2",
    "sqlalchemy>=2.0.46",
    "t-tech-investments>=0.3.3",
    "ta-lib>=0.6.8",
    "uuid6>=2025.0.1",
    "yfinance>=1.2.0",
]

[dependency-groups]
dev = [
    "types-pyyaml>=6.0.12.20250915",
]

[tool.ruff]
line-length = 88
target-version = "py313"
exclude = ["migrations/*", "tests/*"]
fix = true

[[tool.uv.index]]
url = "https://opensource.tbank.ru/api/v4/projects/238/packages/pypi/simple"
default = true

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E203"]
preview = true

[tool.ruff.format]
indent-style = "space"
quote-style = "double"
skip-magic-trailing-comma = false
docstring-code-format = true
docstring-code-line-length = 88

[tool.mypy]
ignore_missing_imports = true
explicit_package_bases = true 
namespace_packages = true
